Rename catalog_entry_form.js to CatalogEntryForm.js (React component naming convention)
